Added a lot of maps functionality

+++ app/src/main/java/com/lambton/projects/tovisit_chaitanya_c0777253_android/viewmodels/FavouritePlaceViewModel.java
@@ -6,6 +6,8 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
 import androidx.lifecycle.LiveData;
 
+import com.google.android.gms.maps.model.Polyline;
+import com.lambton.projects.tovisit_chaitanya_c0777253_android.callbacks.Callbacks;
 import com.lambton.projects.tovisit_chaitanya_c0777253_android.models.FavouritePlace;
 import com.lambton.projects.tovisit_chaitanya_c0777253_android.repositories.FavouritePlaceRepository;
 
@@ -24,18 +26,23 @@ public class FavouritePlaceViewModel extends AndroidViewModel
         System.out.println(3);
     }
 
-    public void insert(FavouritePlace favouritePlace)
+    public void insert(FavouritePlace favouritePlace, Callbacks callbacks)
     {
-        mRepository.insert(favouritePlace);
+        mRepository.insert(favouritePlace, callbacks);
     }
 
     public void update(FavouritePlace favouritePlace)
     {
+        System.out.println("Updating title: "+favouritePlace.getTitle()+" visited: "+favouritePlace.isVisited());
         mRepository.update(favouritePlace);
     }
 
     public void delete(FavouritePlace favouritePlace)
     {
+        if(favouritePlace == null)
+        {
+            return;
+        }
         mRepository.delete(favouritePlace);
     }
